requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
chromadb==0.6.3
numpy==1.26.4
pypdf==5.1.0
python-docx==1.1.2
python-multipart==0.0.20
aiofiles==24.1.0